==> 02_cuadernos/01_filtrado y estraccion.ipynb <==
# Librerias

import pickle
import pandas as pd
import numpy as np
import warnings
import re
from unidecode import unidecode
from tqdm import tqdm

import requests
from bs4 import BeautifulSoup

warnings.filterwarnings('ignore')

# Mis funciones

import import_ipynb
from mis_funciones import *

# Extracción de urls y otros datos para la extracción masiva

urls_shows = []
locales_shows = []
nombre_shows = []
n_opiniones = []  # Lista para almacenar el número de opiniones

for i in range(1, 10):
    url = f'https://www.atrapalo.com/entradas/madrid/teatro-y-danza/monologos/p-{i}/'
    
    res1 = requests.get(url)
    sopa = BeautifulSoup(res1.content, 'html.parser')

    urls = sopa.find_all("a", {"class": "product-name"})
    for url_evento in urls:
        urls_shows.append(url_evento['href'])
        nombre_local = url_evento.find_next("a", {"class": "locality GATrackEvent_ubicacion hide-for-small-only"}).text
        locales_shows.append(nombre_local)

        nombre_show = url_evento.text.strip()
        nombre_shows.append(nombre_show)
        
        # Obtener el número de opiniones
        url_completa = 'https://www.atrapalo.com/' + url_evento['href']
        response = requests.get(url_completa)
        soup = BeautifulSoup(response.content, 'html.parser')

        # Buscar el texto que contiene el número de opiniones
        texto_opiniones = soup.get_text()

        # Buscar un patrón de texto que incluya el número de opiniones
        match = re.search(r'(\d+)\s+opiniones', texto_opiniones)

        if match:
            # Extraer el número de opiniones
            num_opiniones = int(match.group(1))
            n_opiniones.append(num_opiniones)
        else:
            # Si no se encuentra el texto, agregar un valor predeterminado (por ejemplo, -1)
            n_opiniones.append(-1)



# Crear el DataFrame
df = pd.DataFrame({'url': urls_shows, 'local': locales_shows, 'nombre_show': nombre_shows, 'n_opiniones': n_opiniones})

df

Necesito unicamente los que aparecen en el siguiente data frame:

df_cuentas_locales = pd.read_csv("../01_data/cuentas_locales.csv", delimiter=",", header=0)
df_cuentas_locales

Y los shows que tengan df_show['local'] = df_cuentas_locales['local']

# Filtrar el DataFrame df para mantener solo las filas que coincidan con los locales en df_cuentas_locales['local']
df_filtrado = df[df['local'].isin(df_cuentas_locales['local'])]

# Añadir el nombre del espectáculo usando un bucle
df_filtrado['nombre_show'] = ''

for index, row in df_filtrado.iterrows():
    # Obtener la URL de la fila actual
    url = row['url']
    
    # Extraer el nombre del espectáculo de la URL
    nombre_show = url.split('/')[-2].replace('-', ' ').title()  # Obtener el nombre del espectáculo desde la URL
    
    # Limpiar el nombre del espectáculo (eliminar todo después de '_')
    nombre_show = nombre_show.split('_')[0] 
    
    # Asignar el nombre del espectáculo a la columna correspondiente
    df_filtrado.at[index, 'nombre_show'] = nombre_show

# Imprimir el DataFrame filtrado con el nombre del espectáculo
df_filtrado

# Calcular el número de páginas (son 10 comentarios por página)
df_filtrado['num_paginas'] = (df_filtrado['n_opiniones'] // 10) + 1
df_filtrado

# Cambio los'-1' por 0
df_filtrado['n_opiniones'] = df_filtrado['n_opiniones'].replace(-1, 0)
df_filtrado

df_filtrado.head(20)

# Guardamos el data frame de filtrado para el scraping masivo

# Especifico la ruta y el nombre del archivo CSV
ruta_archivo_csv = '../01_data/cuentas_scraping.csv'

# Guardo el DataFrame en un archivo CSV
df_filtrado.to_csv(ruta_archivo_csv, index=False)

# Extracción masiva con limpieza y guardado de datos

# Función para extraer comentarios de una página específica
def extraer_comentarios(url, nombre_show, nombre_local):
    """
    Esta función extrae comentarios de una página web y devuelve una lista de diccionarios con información detallada sobre cada comentario.
    
    Requiere las siguientes bibliotecas:
    - requests
    - BeautifulSoup (bs4)
    """
    response = requests.get(url)
    soup = BeautifulSoup(response.content, "html.parser")
    comentarios = soup.select(".rws-results-review-item")
    lista_comentarios = []

    for comentario in comentarios:
        nota_total = comentario.select_one(".aui-ratingBadge--large").text
        fecha = comentario.select_one(".meta-info__date").text
        grupo_de = comentario.select_one(".meta-info__company .aui-badge-text").text

        aspectos = comentario.select(".rws-circle-chart-list__item")

        if aspectos:
            calidad = aspectos[0].select_one(".rws-circle-chart__value-container span").text
            puesta_en_escena = aspectos[1].select_one(".rws-circle-chart__value-container span").text
            interpretacion = aspectos[2].select_one(".rws-circle-chart__value-container span").text

            comentario_dict = {
                "fecha": fecha,
                "nota_total": nota_total,
                "grupo_de": grupo_de,
                "calidad_del_espectáculo": calidad,
                "puesta_en_escena": puesta_en_escena,
                "interpretación_artística": interpretacion,
                "nombre_show": nombre_show,  
                "nombre_local": nombre_local  
            }

            lista_comentarios.append(comentario_dict)

    return lista_comentarios

def convertir_fecha(fecha):
    meses = {
        "enero": "January",
        "febrero": "February",
        "marzo": "March",
        "abril": "April",
        "mayo": "May",
        "junio": "June",
        "julio": "July",
        "agosto": "August",
        "septiembre": "September",
        "octubre": "October",
        "noviembre": "November",
        "diciembre": "December"
    }
    partes = fecha.split()
    partes[1] = meses[unidecode(partes[1].lower())]
    fecha_nueva = " ".join(partes)
    return pd.to_datetime(fecha_nueva, format='%d %B %Y', errors='coerce')

## 1er día

# URL base de la página
base_url = "https://www.atrapalo.com"

# Inicializar lista para almacenar todos los comentarios
todos_los_comentarios = []

# Iterar a través de las filas de df_filtrado
for index, row in df_filtrado.iterrows():
    url = f"{base_url}{row['url']}"
    nombre_show = row['nombre_show']
    nombre_local = row['local']
    num_paginas = row['num_paginas']
    
    # Iterar a través de las páginas
    for i in range(1, num_paginas + 1):
        url_pagina = f"{url}page-{i}/"
        comentarios_pagina = extraer_comentarios(url_pagina, nombre_show, nombre_local)
        todos_los_comentarios.extend(comentarios_pagina)
    
    # Pasar a DataFrame los datos
    df_comentarios = pd.DataFrame(todos_los_comentarios)

    # Aplicar la función a la columna 'fecha'
    df_comentarios['fecha'] = df_comentarios['fecha'].apply(convertir_fecha)

    # Quitar 'sobre 10' de las columnas de calidad del espectáculo, puesta en escena e interpretación artística
    df_comentarios['calidad_del_espectáculo'] = df_comentarios['calidad_del_espectáculo'].str.replace(r'\s+sobre\s+10', '', regex=True)
    df_comentarios['puesta_en_escena'] = df_comentarios['puesta_en_escena'].str.replace(r'\s+sobre\s+10', '', regex=True)
    df_comentarios['interpretación_artística'] = df_comentarios['interpretación_artística'].str.replace(r'\s+sobre\s+10', '', regex=True)

    # Guardar el DataFrame en un archivo CSV
    df_comentarios.to_csv(f"../01_data/comentarios_{nombre_show}.csv", index=False)
    
# Se ha quedado en la chocita

## 2º día

En la anterior linea de código se quedo en 'Monologos De Humor Chocita Del Loro Gran Via'.

# Volvemos a cargar el archivo de filtrado
df_filtrado = pd.read_csv("../01_data/cuentas_scraping.csv", delimiter=",", header=0)
df_filtrado

Hemos observado que hay cierto número de opiniones que se han scrapeado mal y tienen un '10' extra delante. Lo limpiamos:

# Iterar sobre las filas del DataFrame
for index, row in df_filtrado.iterrows():
    # Convertir n_opiniones a string para verificar si comienza con '10'
    n_opiniones = str(row['n_opiniones'])
    
    # Verificar si comienza con '10'
    if n_opiniones.startswith('10'):
        # Quitar '10' del inicio
        df_filtrado.at[index, 'n_opiniones'] = int(n_opiniones[2:])

# Ver el DataFrame actualizado
df_filtrado

Como 'La chocita del loro' dío problemas seguimos por el siguiente regestro. Para ello buscamos su indice:

indice_reanudacion = df_filtrado[df_filtrado['nombre_show'] == 'Impro Monologos Clandestinos'].index[0]
indice_reanudacion

Reanudamos el scraping:

# URL base de la página
base_url = "https://www.atrapalo.com"

# Inicializar lista para almacenar todos los comentarios
todos_los_comentarios = []

# Iterar a través de las filas de df_filtrado desde el índice de reanudación
for index, row in df_filtrado.iloc[indice_reanudacion:].iterrows():
    url = f"{base_url}{row['url']}"
    nombre_show = row['nombre_show']
    nombre_local = row['local']
    num_paginas = row['num_paginas']
    
    # Iterar a través de las páginas
    for i in range(1, num_paginas + 1):
        url_pagina = f"{url}page-{i}/"
        comentarios_pagina = extraer_comentarios(url_pagina, nombre_show, nombre_local)
        todos_los_comentarios.extend(comentarios_pagina)
    
    # Pasar a DataFrame los datos
    df_comentarios = pd.DataFrame(todos_los_comentarios)

    # Aplicar la función a la columna 'fecha'
    df_comentarios['fecha'] = df_comentarios['fecha'].apply(convertir_fecha)

    # Quitar 'sobre 10' de las columnas de calidad del espectáculo, puesta en escena e interpretación artística
    df_comentarios['calidad_del_espectáculo'] = df_comentarios['calidad_del_espectáculo'].str.replace(r'\s+sobre\s+10', '', regex=True)
    df_comentarios['puesta_en_escena'] = df_comentarios['puesta_en_escena'].str.replace(r'\s+sobre\s+10', '', regex=True)
    df_comentarios['interpretación_artística'] = df_comentarios['interpretación_artística'].str.replace(r'\s+sobre\s+10', '', regex=True)

    # Guardar el DataFrame en un archivo CSV
    df_comentarios.to_csv(f"../01_data/comentarios_{nombre_show}.csv", index=False)

# Unimos los csv y creamos una sola BBDD

# Volvemos a cargar el archivo de filtrado
df1 = pd.read_csv("../01_data/comentarios_Situaciones Incomodas Marcos Arizmendi.csv", delimiter=",", header=0)
df2 = pd.read_csv("../01_data/comentarios_Histericas.csv", delimiter=",", header=0)
df_comentarios = pd.concat([df1, df2], ignore_index=True)
df_comentarios

# Guardo el DataFrame en un archivo CSV
df_comentarios.to_csv('../01_data/comentarios_atrapalo.csv', index=False)
