==> 02_cuadernos/04_eda.ipynb <==
# Librerias

import pandas as pd
import sweetviz as sv

# Cargamos datos

# Cargar el archivo CSV
df_comentarios = pd.read_csv('../01_data/comentarios_atrapalo.csv', delimiter=",", header=0)
df_comentarios

# Cargar el archivo CSV
df_beer1 = pd.read_csv('../01_data/comentarios_YO NUNCA SHOW,  de Raúl Massana.csv', delimiter=",", header=0)
df_beer2 = pd.read_csv('../01_data/comentarios_Quique_Matilla__Vastante_vien.csv', delimiter=",", header=0)

# Unimos archivos

df_beer = pd.concat([df_beer1, df_beer2], ignore_index=True)
df_beer

df_comentarios_final = pd.concat([df_comentarios, df_beer], ignore_index=True)
df_comentarios_final

# Eliminamos los datos que al final no vamos a usar

chocita_rows = df_comentarios_final[df_comentarios_final['nombre_local'].str.contains('Chocita', case=False)]
chocita_rows

df_comentarios_final = df_comentarios_final[df_comentarios_final['nombre_local'] != 'Teatro Chocita del Loro - Gran Vía']
df_comentarios_final

# Guardamos BBDD final

# Guardo el DataFrame en un archivo CSV
df_comentarios_final.to_csv('../01_data/total_comentarios.csv', index=False)

df_tipo_show = pd.read_csv('../01_data/shows_atrapalo.csv', delimiter=",", header=0)

# Unir los DataFrames usando 'nombre_show' como clave de unión
df_final = pd.merge(df_comentarios_final, df_tipo_show[['nombre_show', 'tipo_show']], on='nombre_show', how='left')

# Llenar los valores nulos con 'show'
df_final['tipo_show'].fillna('show', inplace=True)
df_final

# Convertir las columnas 'nombre_show' a minúsculas en ambos DataFrames
df_comentarios_final['nombre_show'] = df_comentarios_final['nombre_show'].str.lower()
df_tipo_show['nombre_show'] = df_tipo_show['nombre_show'].str.lower()

# Unir los DataFrames usando 'nombre_show' como clave de unión
df_final = pd.merge(df_comentarios_final, df_tipo_show[['nombre_show', 'tipo_show']], on='nombre_show', how='left')

# Llenar los valores nulos con 'show'
df_final['tipo_show'].fillna('show', inplace=True)
df_final

# Procedemos al EDA

report = sv.analyze(df_final)
report.show_html()
