==> 02_cuadernos/06_filtrado_cuentas_locales.ipynb <==
Como por el momento únicamente analizaremos las cuentas de los locales filtramos para quedarnos solamente con esos datos

# Librerias

import pandas as pd

# Abrimos archivos

# Cargar el archivo CSV
df_comentarios_final  = pd.read_csv('../01_data/total_comentarios.csv', delimiter=",", header=0)

df_comentarios_final

df_cuentas_locales = pd.read_csv("../01_data/cuentas_locales.csv", delimiter=",", header=0)
df_cuentas_locales

import difflib

# Crear una lista para almacenar los nombres y sus coincidencias aproximadas
coincidencias = []

# Iterar sobre los nombres en df_cuentas_locales
for nombre_cuentas_locales in df_cuentas_locales['nombre_show']:
    # Encontrar la mejor coincidencia aproximada en df_comentarios_final
    coincidencia_aproximada = difflib.get_close_matches(nombre_cuentas_locales, df_comentarios_final['nombre_show'], n=1)
    
    # Si se encontró una coincidencia, añadirla a la lista
    if coincidencia_aproximada:
        coincidencias.append((nombre_cuentas_locales, coincidencia_aproximada[0]))

# Crear un DataFrame con las coincidencias
df_coincidencias = pd.DataFrame(coincidencias, columns=['Nombre en Cuentas Locales', 'Nombre en Comentarios Finales'])

# Mostrar el DataFrame
df_coincidencias

# Filtrar las filas de df_comentarios_final donde el nombre_show coincide con las coincidencias
df_comentarios_cuentas_locales = df_comentarios_final[df_comentarios_final['nombre_show'].isin(df_coincidencias['Nombre en Comentarios Finales'])]

# Reiniciar los índices si es necesario
df_comentarios_cuentas_locales.reset_index(drop=True, inplace=True)

# Mostrar el DataFrame
df_comentarios_cuentas_locales

# Guardo el DataFrame en un archivo CSV
df_comentarios_cuentas_locales.to_csv('../01_data/comentarios_cuentas_locales.csv', index=False)
